A little perspiration

+++ 0x01-variables_if_else_while/9-print_comb.c
@@ -3,12 +3,12 @@
 /**
  * main - Holds Code
  * Description: Ascii 48 to 57 is decimal 0 to 9
- * Working with a nested for loop
+ * Conditional statement in for loop
  * Return: retuns 0
  */
 int main(void)
 {
-	short x;
+	int x;
 
 	for (x = 48; x < 58; x++)
 	{
